rough draft of components and redux development, nothing working so far
