+++ app/reducers/quizReducers.js
@@ -1,15 +1,20 @@
-import { NEW_QUESTIONS } from 'actions/article'; // TODO: create new action here
+import { NEW_QUESTION, SWITCH_ACTIVE_QUESTION } from 'actions/article'; // TODO: create new action here
 import tmpQuestions from '../assets/tmpQuestions.json';
+// import React Update Addon to add new elements to the array
+// to avoid the deep merging issue
 
 const initialState = {
-  questions: tmpQuestions.initialQuestion
+  questions: [tmpQuestions.initialQuestion], 
+  activeQuestion: 0
 }
 
 export default function quizReducer(state = initialState, action) {
-  switch (action.type) {
-    case NEW_QUESTIONS:
-      return Object.assign({}, state, { questions: action.questions });
-    default:
-      return state;
-  }
+    switch (action.type) {
+        case NEW_QUESTION:
+            return update(state, {{$add}})
+        case SWITCH_ACTIVE_QUESTION:
+            return Object.assign({}, state, { activeQuestion: 0})
+        default:
+            return state;
+    }
 }
